DB410c: DB820c: mark Linux kernel package as held
By default, hold the kernel debian package. To work properly, the kernel Image
and the kernel modules must be in sync. On Dragonboards, the kernel Image is
stored in the boot partition as an Android boot image, while the kernel modules
reside in the rootfs in /lib/modules.

When the kernel package is upgraded, new modules are installed in the rootfs,
however the kernel Image in the boot partition is not updating, which can lead
to kernel/module mismatch.

Until we implement a better option, we hold the kernel package so that it is not
ugpraded. We will provide instruction to un-hold the package for power users
that understand the risks they are taking.

Change-Id: Ic18e753ea33e1183d3655ed00d2fe7d5a9370c65
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>